+++ demisto_sdk/commands/lint/linter.py
@@ -1,4 +1,5 @@
 # STD python packages
+import copy
 import hashlib
 import io
 import json
@@ -74,7 +75,8 @@ class Linter:
             "is_long_running": False,
             "lint_unittest_files": [],
             "additional_requirements": [],
-            "docker_engine": docker_engine
+            "docker_engine": docker_engine,
+            "is_script": False,
         }
         # Pack lint status object - visualize it
         self._pkg_lint_status: Dict = {
@@ -188,7 +190,7 @@ class Linter:
             yml_obj: Dict = YAML().load(yml_file)
             if isinstance(yml_obj, dict):
                 script_obj = yml_obj.get('script', {}) if isinstance(yml_obj.get('script'), dict) else yml_obj
-
+            self._facts['is_script'] = True if 'Scripts' in yml_file.parts else False
             self._facts['is_long_running'] = script_obj.get('longRunning')
             self._pkg_lint_status["pack_type"] = script_obj.get('type')
         except (FileNotFoundError, IOError, KeyError):
@@ -295,15 +297,20 @@ class Linter:
         warning = []
         error = []
         other = []
-        if self._facts["lint_files"]:
-            exit_code: int = 0
-            for lint_check in ["flake8", "XSOAR_linter", "bandit", "mypy", "vulture"]:
-                exit_code = SUCCESS
-                output = ""
+        exit_code: int = 0
+        for lint_check in ["flake8", "XSOAR_linter", "bandit", "mypy", "vulture"]:
+            exit_code = SUCCESS
+            output = ""
+            if self._facts["lint_files"] or self._facts["lint_unittest_files"]:
                 if lint_check == "flake8" and not no_flake8:
-                    exit_code, output = self._run_flake8(py_num=self._facts["python_version"],
-                                                         lint_files=self._facts["lint_files"])
-                elif lint_check == "XSOAR_linter" and not no_xsoar_linter:
+                    flake8_lint_files = copy.deepcopy(self._facts["lint_files"])
+                    # if there are unittest.py then we would run flake8 on them too.
+                    if self._facts['lint_unittest_files']:
+                        flake8_lint_files.extend(self._facts['lint_unittest_files'])
+                    exit_code, output = self._run_flake8(py_num=self._facts["python_version"], lint_files=flake8_lint_files)
+
+            if self._facts["lint_files"]:
+                if lint_check == "XSOAR_linter" and not no_xsoar_linter:
                     exit_code, output = self._run_xsoar_linter(py_num=self._facts["python_version"],
                                                                lint_files=self._facts["lint_files"])
                 elif lint_check == "bandit" and not no_bandit:
@@ -315,40 +322,20 @@ class Linter:
                     exit_code, output = self._run_vulture(py_num=self._facts["python_version"],
                                                           lint_files=self._facts["lint_files"])
 
-                # check for any exit code other than 0
-                if exit_code:
-                    error, warning, other = split_warnings_errors(output)
-                if exit_code and warning:
-                    self._pkg_lint_status["warning_code"] |= EXIT_CODES[lint_check]
-                    self._pkg_lint_status[f"{lint_check}_warnings"] = "\n".join(warning)
-                if exit_code & FAIL:
-                    self._pkg_lint_status["exit_code"] |= EXIT_CODES[lint_check]
-                    # if the error were extracted correctly as they start with E
-                    if error:
-                        self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(error)
-                    # if there were errors but they do not start with E
-                    else:
-                        self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(other)
-
-        if self._facts['lint_unittest_files']:
-            for lint_check in ["flake8"]:
-                exit_code = SUCCESS
-                output = ""
-                if lint_check == "flake8" and not no_flake8:
-                    exit_code, output = self._run_flake8(py_num=self._facts["python_version"],
-                                                         lint_files=self._facts["lint_unittest_files"])
-                if exit_code:
-                    error, warning, other = split_warnings_errors(output)
-                if exit_code & FAIL:
-                    self._pkg_lint_status["exit_code"] |= EXIT_CODES[lint_check]
-                    if error:
-                        self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(error)
-                    # for errors whihc start with 'E'
-                    else:
-                        self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(other)
-                if exit_code and WARNING:
-                    self._pkg_lint_status["warning_code"] |= EXIT_CODES[lint_check]
-                    self._pkg_lint_status[f"{lint_check}_warnings"] = "\n".join(warning)
+            # check for any exit code other than 0
+            if exit_code:
+                error, warning, other = split_warnings_errors(output)
+            if exit_code and warning:
+                self._pkg_lint_status["warning_code"] |= EXIT_CODES[lint_check]
+                self._pkg_lint_status[f"{lint_check}_warnings"] = "\n".join(warning)
+            if exit_code & FAIL:
+                self._pkg_lint_status["exit_code"] |= EXIT_CODES[lint_check]
+                # if the error were extracted correctly as they start with E
+                if error:
+                    self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(error)
+                # if there were errors but they do not start with E
+                else:
+                    self._pkg_lint_status[f"{lint_check}_errors"] = "\n".join(other)
 
     def _run_flake8(self, py_num: float, lint_files: List[Path]) -> Tuple[int, str]:
         """ Runs flake8 in pack dir
@@ -403,6 +390,7 @@ class Linter:
                 myenv['LONGRUNNING'] = 'True'
             if py_num < 3:
                 myenv['PY2'] = 'True'
+            myenv['is_script'] = str(self._facts['is_script'])
             stdout, stderr, exit_code = run_command_os(
                 command=build_xsoar_linter_command(lint_files, py_num, self._facts.get('support_level', 'base')),
                 cwd=self._pack_abs_dir, env=myenv)