+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added new validation of indicators usage in CommandResults to `XSOAR-linter`.
 * Running **demisto-sdk lint** will automatically run on changed files (same behavior as the -g flag).
 * Removed supported version message from the documentation when running **generate_docs**.
 * Added a print to indicate backwards compatibility is being checked in **validate** command.
@@ -8,7 +9,14 @@
 * Fixed an issue where the **validate** command did not create the *id_set.json* file when running with *-a* flag.
 * Added price change validation in the **validate** command.
 * Added validations that checks in read-me for empty sections or leftovers from the auto generated read-me that should be changed.
-
+* Added new code validation for *NotImplementedError* to raise a warning in `XSOAR-linter`.
+* Added validation for support types in the pack metadata file.
+* Added support for *--template* flag in **demisto-sdk init** command.
+* Fixed an issue with running **validate** on master branch where the changed files weren't compared to previous commit when using the *-g* flag.
+* Fixed an issue where the `XSOAR-linter` ran *NotImplementedError* validation on scripts.
+* Added support for Auto-Extract feature validation in incident types in the **validate** command.
+* Fixed an issue in the **lint** command where the *-i* flag was ignored.
+* Fixed an issue in the **lint** command where flake8 ran twice.
 
 # 1.2.12
 * Bandit now reports also on medium severity issues.